fix(Versions): remove code specific to v1
